+++ src/helper.cpp
@@ -28,6 +28,26 @@ const QString Helper::logType[ LOG_TYPE_COUNT ] =
     "Ignored.txt",
 };
 
+const QList<qint32> Helper::blueCodedList =
+{
+    1004, 1024, 1043, 1046, 1052, 1054, 1055, 1062, 1068, 1072, 1099,
+    1112, 1120, 1123, 1125, 1138, 1163, 1166, 1170, 1172, 1173, 1189,
+    1204, 1210, 1217, 1275, 1292, 1307, 1308, 1312, 1332, 1338, 1367,
+    1369, 1370, 1520, 1547, 1551, 1565, 1600, 1607, 1611, 1656, 1675,
+    1681, 1695, 1706, 1715, 1751, 1754, 1840, 1965, 2003, 2058, 2062,
+    2144, 2205, 2217, 2264, 2268, 2359, 1008, 1017, 1051, 1054, 1082,
+    1099, 1104, 1105, 1181, 1199, 1222, 1279, 1343, 1358, 1388, 1456,
+    1528, 1677, 1773, 1777, 1778, 1780, 1796, 1799, 2156, 2167, 2241,
+    2248, 2362, 2421, 1098, 1220, 1264, 1342, 1361, 1823, 2302, 2488,
+    2585, 2372, 1492, 1576, 1100, 1347, 1050, 1015, 1666, 1745, 2043,
+    1200, 2628, 1016, 1739, 1853, 2708, 2757, 1498, 2448, 2801, 1031,
+    1265, 1414, 1420, 1429, 1214, 1489, 1707, 2543, 1101, 1283, 1604,
+    1428, 2707, 1023, 1069, 1071, 1132, 1286, 1854, 2910, 1005, 2682,
+    1348, 2615, 2617, 1884, 1169, 1540, 1645, 1939, 1179, 3053, 1803,
+    2377, 1000, 1021, 1500, 1501, 1515, 1547, 1803, 2377, 3111, 3202,
+    3191, 3149, 3,
+};
+
 QInputDialog* Helper::createInputDialog(QWidget* parent, QString& label,
                                         QInputDialog::InputMode mode,
                                         int width, int height)
@@ -231,6 +251,11 @@ qint32 Helper::serNumtoInt(QString& sernum)
         sernum_i = strToInt( sernum, 10 );
 
     return sernum_i;
+}
+
+bool Helper::isBlueCodedSerNum(const quint32& sernum)
+{
+    return blueCodedList.contains( sernum );
 }
 
 void Helper::logToFile(const LogTypes& type, const QString& text,
@@ -441,8 +466,7 @@ QHostAddress Helper::getPrivateIP()
     QList<QHostAddress> ipList = QNetworkInterface::allAddresses();
 
     //Default to our localhost address if nothing valid is found.
-    QHostAddress ipAddress{ QHostAddress::LocalHost };
-
+    QHostAddress ipAddress{ QHostAddress::Null };
     for ( int i = 0; i < ipList.size(); ++i )
     {
         QString tmp = ipList.at( i ).toString();